+++ easygraphics/graphwin.py
@@ -48,6 +48,7 @@ class GraphWin(QtWidgets.QWidget):
         self._frames_to_skip_count = 0
         self._last_fps_time = 0
         self._frames_skipped = 0
+        self._wait_close = False
 
     def get_width(self):
         return self._width
@@ -130,18 +131,40 @@ class GraphWin(QtWidgets.QWidget):
         pause and wait for mouse click or keyboard hit
         """
         self.real_update()
+        if self._wait_close:
+            self._do_close()
+            return False
         self._wait_event.clear()
         self._wait_event.wait()
 
-    def closeEvent(self, e: QtGui.QCloseEvent):
+    def _do_close(self):
         self.close_signal.emit()
         self._is_run = False
         self._canvas.close()
-        self._wait_event.set()
-        self._mouse_event.set()
-        self._key_event.set()
-        self._char_key_event.set()
+        self.close()
         self._app.quit()
+
+    def closeEvent(self, e: QtGui.QCloseEvent):
+        if self._wait_event is not None:
+            self._wait_event.set()
+            self._wait_event = None
+        if self._mouse_event is not None:
+            self._mouse_event.set()
+            self._mouse_event = None
+        if self._key_event is not None:
+            self._key_event.set()
+            self._key_event = None
+        if self._char_key_event is not None:
+            self._char_key_event.set()
+            self._char_key_event = None
+        if self._immediate:
+            self._do_close()
+        else:
+            self._is_run = False
+            self._wait_close = True
+
+    def is_quitting(self):
+        return self._wait_close
 
     def is_run(self) -> bool:
         return self._is_run
@@ -163,20 +186,26 @@ class GraphWin(QtWidgets.QWidget):
         """
         if self._immediate:
             raise RuntimeError("Must set render mode to MANUAL to use delay()!")
+        elif self._wait_close:
+            self._do_close()
         nanotime = milliseconds * 1000000
         start_wait_time = time.perf_counter_ns()
         self.real_update()
         while time.perf_counter_ns() - start_wait_time < nanotime:
             time.perf_counter_ns()
 
-    def delay_fps(self, fps: int):
+    def delay_fps(self, fps: int) -> bool:
         """
         Delay to control fps without frame skipping. Never skip frames.
 
         :param fps: the desire fps
+        :return: False the graphics window is closed. True otherwise.
         """
         if self._immediate:
             raise RuntimeError("Must set render mode to MANUAL to use delay()!")
+        elif self._wait_close:
+            self._do_close()
+            return False
         nanotime = 1000000000 // fps
         if self._last_fps_time == 0:
             self._last_fps_time = time.perf_counter_ns()
@@ -185,6 +214,7 @@ class GraphWin(QtWidgets.QWidget):
         while tt - self._last_fps_time < nanotime:
             tt = time.perf_counter_ns()
         self._last_fps_time = tt
+        return True
 
     def delay_jfps(self, fps: int, max_skip_count: int = 10) -> bool:
         """
@@ -198,6 +228,9 @@ class GraphWin(QtWidgets.QWidget):
         """
         if self._immediate:
             raise RuntimeError("Must set render mode to MANUAL to use delay()!")
+        elif self._wait_close:
+            self._do_close()
+            return False
         nanotime = 1000000000 // fps
         if self._frames_to_skip_count > 0:
             self._frames_to_skip_count -= 1
@@ -241,6 +274,9 @@ class GraphWin(QtWidgets.QWidget):
         self.real_update()
         if nt - self._key_char_msg.get_time() > 100000000:
             # if the last char msg is 100ms ago, we wait for a new msg
+            if self._wait_close:
+                self._do_close()
+                return ' '
             self._char_key_event.clear()
             self._char_key_event.wait()
         if not self._is_run:
@@ -262,6 +298,9 @@ class GraphWin(QtWidgets.QWidget):
         self.real_update()
         if nt - self._key_msg.get_time() > 100000000:
             # if the last key msg is 100ms ago, we wait for a new msg
+            if self._wait_close:
+                self._do_close()
+                return QtCore.Qt.Key_Escape, QtCore.Qt.NoModifier
             self._key_event.clear()
             self._key_event.wait()
         if not self._is_run:
@@ -284,6 +323,9 @@ class GraphWin(QtWidgets.QWidget):
         self.real_update()
         if nt - self._mouse_msg.get_time() > 100000000:
             # if the last key msg is 100ms ago, we wait for a new msg
+            if self._wait_close:
+                self._do_close()
+                return 0, 0, 0, QtCore.Qt.NoButton
             self._mouse_event.clear()
             self._mouse_event.wait()
         if not self._is_run: