+++ examples/tutorials/13_cursor_pos.py
@@ -4,9 +4,9 @@ init_graph(800, 600)
 set_render_mode(RenderMode.RENDER_MANUAL)
 
 while is_run():
-    x, y = get_cursor_pos()
-    clear_device()
-    draw_text(0, 600, "%d,%d" % (x, y))
-    delay_fps(30)
+    if delay_fps(30):
+        x, y = get_cursor_pos()
+        clear_device()
+        draw_text(0, 600, "%d,%d" % (x, y))
 
 close_graph()