+++ examples/tutorials/13_input_game.py
@@ -56,7 +56,9 @@ if __name__ == "__main__":
                     show_goodbye()
                     close_graph()
                     exit()
-            delay_fps(60)
-            fill_rect(x - 2, y - 22, x + 22, y + 2)  # clear the char
+            if delay_fps(60):
+                fill_rect(x - 2, y - 22, x + 22, y + 2)  # clear the char
+            else:
+                break
 
     close_graph()