+++ examples/tutorials/14_change_color.py
@@ -5,11 +5,11 @@ init_graph(600, 400)
 set_render_mode(RenderMode.RENDER_MANUAL)
 
 while is_run():
-    if has_mouse_msg():
-        x, y, type, buttons = get_mouse_msg()
-        if type == MouseMessageType.PRESS_MESSAGE:
-            color = get_color(get_background_color())
-            set_background_color(color)
-    delay_fps(60)
+    if delay_fps(60):
+        if has_mouse_msg():
+            x, y, type, buttons = get_mouse_msg()
+            if type == MouseMessageType.PRESS_MESSAGE:
+                color = get_color(get_background_color())
+                set_background_color(color)
 
 close_graph()