+++ examples/turtle/sun.py
@@ -8,7 +8,7 @@ set_color("red")
 set_fill_color("yellow")
 set_speed(100)
 begin_fill()
-while True:
+while is_run():
     forward(200)
     left_turn(170)
     if math.hypot(get_x(), get_y()) < 1: